Selection Tool
Added a drag selection. It doesn't actually select anything yet,
and there are also some conflicts with fills and stroke colours,
but those can be fixed throughout the code. Whenever we draw something
I think we should update the fill() and the stroke() so that shit
doesn't get fucked up.